Move ResultType; introduce ResultFormat without extraneous information
