+++ impl/src/main/scala/quasar/impl/datasource/local/LocalDestination.scala
@@ -18,10 +18,11 @@ package quasar.impl.datasource.local
 
 import slamdata.Predef.{Stream => _, _}
 
+import quasar.api.destination.ResultType
 import quasar.api.resource.ResourcePath
 import quasar.api.table.TableColumn
 import quasar.concurrent.BlockingContext
-import quasar.connector.{Destination, MonadResourceErr, ResourceError, ResultSink, ResultType}
+import quasar.connector.{Destination, MonadResourceErr, ResourceError, ResultSink}
 
 import cats.effect.{ContextShift, Effect}
 import fs2.{io, Stream}
